requirements.txt
google-genai>=1.0.0
customtkinter>=5.2.0
python-dotenv>=1.0.0
Pillow>=10.0.0
